+++ src/output_asio.h
@@ -18,7 +18,7 @@ bool Out_ASIO_Reset();
 // Adds a stream to be mixed into the ASIO output. It should not be freed until ASIO shuts down.
 void Out_ASIO_AddStream(SDL_AudioStream* stream);
 
-double          Out_ASIO_GetFrequency();
+int             Out_ASIO_GetFrequency();
 SDL_AudioFormat Out_ASIO_GetFormat();
 
 // Returns the size of a single sample in bytes.
